Add use transactions sections helper
